+++ frontend/public/components/Dashboard.vue
@@ -2,60 +2,8 @@
     <div>
         <banner></banner>
         <div v-if="currTab == 'dashboard'" id="dashboard">
-            <div id="inbox">
-                <div v-if="estimating == undefined">
-                    <p class="title">Inbox</p>
-                    <ul>
-                        <li v-for="item in inbox" v-bind:key="item.title">
-                            <div class="dashboard-card">
-                                <p @click="openInboxItem(item)">{{item.description}}: {{item.title}}</p>
-                            </div>
-                        </li>
-                    </ul>
-                </div>
-                <div id="estimate-container" v-if="estimating != undefined">
-                    <p class="title">Estimate Task</p>
-                    <p>{{estimating.response.description}}</p>
-                    <p><input type="number" id="estimate-input"></p>
-                    <button @click="submitEstimate">Submit</button>
-                </div>
-            </div>
-            <div id="calendar">
-                <p class="title">Calendar</p>
-                <div id="calendar-table-container">
-                    <table id="calendar-table" cellspacing="0">
-                        <thead>
-                            <tr>
-                                <th v-for="day in daysOfWeek" v-bind:key="day">{{day}}</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr v-for="row in [...Array(6).keys()]" v-bind:key="row">
-                                <td v-for="col in [...Array(7).keys()]" v-bind:key="col" v-bind:class="[ dayHighlighted[calendarToDate(row,col)] ? dayHighlightedClass : '']">
-                                    <div @click="showDay(row,col)" class="calendar-cell clickable" 
-                                        v-bind:class="[
-                                        (col == 0 || col == 6)? weekendClass : '',(calendarToDate(row,col) < 1 || 
-                                        calendarToDate(row,col) > monthLength)? blankDateClass : '',
-                                        (new Date() > new Date(2020, 4, calendarToDate(row,col), 23) ? pastDateClass : ''),
-                                        workloadClasses(calendarToDate(row,col))]"
-                                    >
-                                        {{dayInMonth(row,col) ? calendarToDate(row,col) :''}}
-                                    </div>
-                                    <div class="date-info" 
-                                    v-if="dayShowing == calendarToDate(row,col) && workload[calendarToDate(row,col)].tasks.length">
-                                        <div class="date-info-inner">
-                                            <p v-for="task in workload[calendarToDate(row,col)].tasks" v-bind:key="task.name">
-                                                {{task.name + ": " + task.time + " hours"}}
-                                            </p>
-                                            <i class="fa fa-times clickable" @click="hideDay"></i>
-                                        </div>
-                                    </div>
-                                </td>
-                            </tr>
-                        </tbody>
-                    </table>
-                </div>
-            </div>
+            <inbox></inbox>
+            <calendar></calendar>
             <div id="taskList">
                 <p class="title">My Tasks</p>
                 <ul>
@@ -185,17 +133,20 @@
     import '../css/style.css'
     import requests from '../services/requests.js';
     import Banner from '../components/Banner.vue';
+    import Calendar from '../components/Calendar.vue';
+    import Inbox from '../components/Inbox.vue';
 
     export default {
         components: {
-            Banner
+            Banner,
+            Calendar,
+            Inbox
         },
         data () {
             return {
                 currTab: "dashboard",
                 monthOffset: 5,
                 monthLength: 31,
-                estimating: undefined,
                 currField: '',
                 fieldTypes: ['text','number','text (long)','date','dropdown','multi-select','slider'],
                 numDropdownFields: 0,
@@ -204,7 +155,6 @@
                 inbox: [],
                 taskList: [],
                 calendarDays: 31,
-                daysOfWeek: ["Su","Mo","Tu","We","Th","Fr","Sa"],
                 weekendClass: "weekend-cell",
                 blankDateClass: "blank-cell",
                 pastDateClass: "past-cell",
@@ -312,16 +262,6 @@
         closeFormPreview() {
             this.formPreview = false;
         },
-        showDay(row,col) {
-            this.dayShowing = this.calendarToDate(row,col);
-        },
-        hideDay() {
-            this.dayShowing = undefined;
-        },
-        isWeekend(day) {
-            let dayOfWeek = this.daysOfWeek[(this.monthStartDayOfWeek + day) % this.daysOfWeek.length];
-            return (dayOfWeek == "Sa" || dayOfWeek == "Su");
-        },
         workloadClasses(day) {
             let hours;
             if (this.workload[day])
@@ -334,23 +274,6 @@
                 return "yellow-day"
             else if (hours < 8)
                 return "green-day";
-        },
-        openInboxItem(item) {
-            this.estimating = item;
-        },
-        submitEstimate() {
-            this.taskList.push(
-                new InboxItem
-                (
-                    this.estimating.response.description,
-                    'Leslie Knope',
-                    new Date(this.estimating.response.due_date),
-                    document.getElementById("estimate-input").value
-                ),
-            )
-            this.inbox.splice(this.inbox.indexOf(this.estimating),1);
-            this.estimating = undefined;
-            this.updateWorkload();
         },
         resizeInput(field) {
             let fieldIndex = this.fields.indexOf(field);
@@ -465,14 +388,4 @@
         }
         }
     }
-
-class InboxItem {
-    constructor(description,from,due,estimate) {
-        this.description = description;
-        this.from = from;
-        this.due = due;
-        this.estimate = estimate;
-        this.danger = false;
-    }
-}
 </script>
